refactor(CoherentCache): Split cache model objects into dedicated files
