+++ src/front/macOS/kDriveCore/ServerBridge/Cache/DTOs/User.swift
@@ -1,0 +1,54 @@
+/*
+ Infomaniak kDrive - Desktop
+ Copyright (C) 2023-2025 Infomaniak Network SA
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License
+ along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import Foundation
+
+public struct User: Identifiable, Hashable, Sendable {
+    public var id: Int32 {
+        dbId
+    }
+
+    public var dbId: Int32
+    public var userId: Int32
+    public var name: String
+    public var email: String
+    public var accounts: [Int32: Account]
+    public var avatar: Data?
+    public var isConnected: Bool
+    public var isStaff: Bool
+
+    public init(
+        dbId: Int32,
+        userId: Int32,
+        name: String,
+        email: String,
+        accounts: [Int32: Account],
+        avatar: Data? = nil,
+        isConnected: Bool,
+        isStaff: Bool
+    ) {
+        self.dbId = dbId
+        self.userId = userId
+        self.name = name
+        self.email = email
+        self.accounts = accounts
+        self.avatar = avatar
+        self.isConnected = isConnected
+        self.isStaff = isStaff
+    }
+}
